# headless/api/documentation/openapi.yaml
openapi: "3.0.0"
info:
  title: "phpBB 4 Headless API"
  version: "1.0.0"
  description: |
    RESTful API layer for phpBB 4.
    Provides all forum functionality via JSON for modern frontend applications.
  contact:
    name: "EBTURK"
servers:
  - url: "/api/v1"
    description: "API v1"
security:
  - BearerAuth: []
paths:
...
  /register/verify/{token}:
    get:
      tags: [Register]
      summary: "Verify email"
      operationId: "verifyEmail"
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Email verified"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT